requirements.txt
absl-py==2.1.0
accelerate==0.34.2
addict==2.4.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
attrdict==2.0.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
comm==0.2.2
cycler==0.12.1
Cython==3.0.11
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
docker-pycreds==0.4.0
einops==0.5.0
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.13.1
flash-attn==2.6.3
flatbuffers==1.12
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.2.0
ftfy==6.2.0
gast==0.4.0
gensim==3.8.3
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.34.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googletrans==3.0.0
grpcio==1.66.0
h11==0.9.0
h2==3.2.0
h5py==3.11.0
hpack==3.0.0
hpargparse==0.14.0
hpman==0.13.0
hstspreload==2024.10.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.24.6
hyperframe==5.2.0
idna==2.10
imageio==2.35.1
importlib_metadata==8.4.0
importlib_resources==6.4.4
ipykernel==6.29.5
ipython==8.12.3
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
llava==0.0.1.dev0
lmdb==1.3.0
loguru==0.7.0
lxml==4.9.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mmcv==0.2.14
mpmath==1.3.0
nbclassic==1.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.0
ninja==1.11.1.1
nlg-eval @ git+https://github.com/Maluuba/nlg-eval.git@2ab4528fad5548315cf61e40c2249fec8c8ad233
nltk==3.9.1
nose==1.3.7
notebook==6.5.4
notebook_shim==0.2.4
numpy==1.22.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
oauthlib==3.2.2
omegaconf==2.2.3
opencv-python==4.6.0.66
opt-einsum==3.3.0
ordereddict==1.1
overrides==7.7.0
packaging==24.1
pandas==1.3.5
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
pathtools==0.1.2
peft==0.12.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
portalocker==2.10.1
prometheus_client==0.20.0
promise==2.3
prompt_toolkit==3.0.47
protobuf==3.19.6
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
reFILE==0.4.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.8.0
rpds-py==0.20.0
rsa==4.9
sacrebleu==2.2.0
safetensors==0.4.4
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.1
Send2Trash==1.8.3
sentence-transformers==3.2.0
sentencepiece==0.1.97
sentry-sdk==2.13.0
setproctitle==1.3.3
shortuuid==1.0.13
six==1.16.0
smart-open==6.0.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
terminado==0.18.1
textaugment==2.0.0
textblob==0.18.0.post0
Theano==1.0.5
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.6.11
tinycss2==1.3.0
tokenizers==0.20.1
torch==2.2.2+cu118
torchaudio==2.2.2+cu118
torchvision==0.17.2+cu118
tornado==6.4.1
tqdm==4.65.0
traitlets==5.14.3
transformers==4.46.2
triton==2.2.0
types-python-dateutil==2.9.0.20240821
typing_extensions==4.9.0
uri-template==1.3.0
urllib3==2.2.2
vidaug @ git+https://github.com/okankop/vidaug@1c1ddf2640fe4a9171267d64ae5e3bd70c24d54a
wandb==0.18.3
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wrapt==1.16.0
xdg==6.0.0
zipp==3.20.1
